Move query param data transform into redux selector
