+++ workspaces/ui-v2/src/components/QueryParametersPanel.tsx
@@ -9,48 +9,12 @@ import {
   ChangedYellowBackground,
   RemovedRedBackground,
 } from '<src>/styles';
-import { IFieldRenderer, IShapeRenderer, JsonLike } from '<src>/types';
+import { QueryParameters } from '<src>/types';
 import { ShapeRenderer } from './ShapeRenderer';
 import { Panel } from './Panel';
 
-type QueryParameters = Record<string, IFieldRenderer>;
-
 type QueryParametersPanelProps = {
   parameters: QueryParameters;
-};
-
-// TODO QPB this should move into redux
-// TODO mvoe this into redux
-export const convertShapeToQueryParameters = (
-  shapes: IShapeRenderer[]
-): QueryParameters => {
-  const queryParameters: QueryParameters = {};
-  if (shapes.length !== 1 || !shapes[0].asObject) {
-    if (shapes.length > 1) {
-      console.error('unexpected format for query parameters');
-    }
-    // otherwise loading
-    return {};
-  }
-
-  for (const field of shapes[0].asObject.fields) {
-    let isArray = field.shapeChoices.findIndex(
-      (choice) => choice.jsonType === JsonLike.ARRAY
-    );
-
-    if (isArray > -1) {
-      field.additionalAttributes = ['multiple'];
-      if (field.shapeChoices.length > 1) {
-        field.shapeChoices.splice(isArray, 1);
-      } else {
-        field.shapeChoices = field.shapeChoices[isArray].asArray!.shapeChoices;
-      }
-    }
-
-    queryParameters[field.name] = field;
-  }
-
-  return queryParameters;
 };
 
 export const QueryParametersPanel: FC<QueryParametersPanelProps> = ({
